Add new genome-specific stratification bed files
Add stratification bed files for complex and compound heterozygous
variant in GIABv3.3.2, Platinum Genomes, and RTG for HG001/NA12878, as
well as for the AJ Trio and Chinese son (HG002 to HG005) for
GIABv3.3.2.  These help to stratify complex and compound heterozygous
variants, which tend to be more difficult to detect.